requirements.txt
Flask==2.3.2
flask-cors==3.0.10
google-genai==0.12.0   
gunicorn==20.1.0
python-dotenv==1.0.0 
